Refactor MonoJuvixToMiniC
